Move the simulator into the main code
Fixes #35
